Move home page sidebar to new component
Since this component is only visual and doesn't contain any logic, it
helps improve the readability of the home page component by extracting
it to a new file.
